CS_AppsStore/requirements.txt
adal==1.2.2
bcrypt==3.1.4
cachetools==3.1.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.1.1
django-allauth==0.40.0
django-bootstrap-form==3.4
django-extensions==2.1.3
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.14.0
idna==2.7
Jinja2==2.10
kubernetes==8.0.1
MarkupSafe==1.1.1
netifaces==0.10.9
oauthlib==3.1.0
psycopg2-binary==2.7.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2018.7
PyYAML==5.1.2
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
websocket-client==0.56.0
